Remove pyelftools completely
