+++ .travis.yml
@@ -95,7 +95,6 @@ matrix:
     # Job 1: OpenMPI
     - env:
       - mpi_type=openmpi
-      - extra_pip=pyelftools
       addons:
         apt:
           packages:
@@ -105,7 +104,6 @@ matrix:
     # Job 2: MPICH
     - env:
       - mpi_type=mpich
-      - extra_pip=
       addons:
         apt:
           packages:
@@ -115,7 +113,6 @@ matrix:
     # Job 3: No MPI
     - env:
       - mpi_type=none
-      - extra_pip=
       addons:
         apt:
           packages:
@@ -123,7 +120,6 @@ matrix:
 
 before_install:
   - pip3 install --user setuptools
-  - pip3 install --user toml $extra_pip
 
 script: dpkg --list | grep binutil && make -j4 && make -j4 -C tests && make check
